refactoring and addressing feedback part 1
