+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/Dstu2Transformers.java
@@ -190,14 +190,6 @@ public final class Dstu2Transformers {
     return maybeDatamart.reference().get();
   }
 
-  /** Return null if the given object is null, otherwise return the converted value. */
-  public static <T, R> R convert(T source, Function<T, R> mapper) {
-    if (source == null) {
-      return null;
-    }
-    return mapper.apply(source);
-  }
-
   /**
    * Return null if the source list is null or empty, otherwise convert the items in the list and
    * return a new one.
@@ -209,15 +201,6 @@ public final class Dstu2Transformers {
     List<R> probablyItems =
         source.stream().map(mapper).filter(Objects::nonNull).collect(Collectors.toList());
     return probablyItems.isEmpty() ? null : probablyItems;
-  }
-
-  /** Filter null items and return null if the result is null or empty. */
-  public static <T> List<T> emptyToNull(List<T> items) {
-    if (isEmpty(items)) {
-      return null;
-    }
-    List<T> filtered = items.stream().filter(Objects::nonNull).collect(Collectors.toList());
-    return filtered.isEmpty() ? null : filtered;
   }
 
   /** Throw a MissingPayload exception if the list does not have at least 1 item. */
@@ -234,34 +217,6 @@ public final class Dstu2Transformers {
       throw new MissingPayload();
     }
     return value;
-  }
-
-  /**
-   * Return the result of the given extractor function if the given object is present. The object
-   * will be passed to the apply method of the extractor function.
-   *
-   * <p>Consider this example:
-   *
-   * <pre>
-   * ifPresent(patient.getGender(), gender -> Patient.Gender.valueOf(gender.value()))
-   * </pre>
-   *
-   * This is equivalent to this standard Java code.
-   *
-   * <pre>
-   * Gender gender = patient.getGender();
-   * if (gender == null) {
-   *   return null;
-   * } else {
-   *   return Patient.Gender.valueOf(gender.value());
-   * }
-   * </pre>
-   */
-  public static <T, R> R ifPresent(T object, Function<T, R> extract) {
-    if (object == null) {
-      return null;
-    }
-    return extract.apply(object);
   }
 
   /**
