+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/organization/Dstu2OrganizationTransformer.java
@@ -1,9 +1,9 @@
 package gov.va.api.health.dataquery.service.controller.organization;
 
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asCodeableConceptWrapping;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.convert;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.emptyToNull;
 import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.convert;
+import static gov.va.api.health.dataquery.service.controller.Transformers.emptyToNull;
 import static java.util.Arrays.asList;
 
 import gov.va.api.health.dataquery.service.controller.EnumSearcher;
