+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/practitioner/Dstu2PractitionerTransformer.java
@@ -2,10 +2,10 @@ package gov.va.api.health.dataquery.service.controller.practitioner;
 
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asCodeableConceptWrapping;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReference;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.convert;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.emptyToNull;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.ifPresent;
 import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.convert;
+import static gov.va.api.health.dataquery.service.controller.Transformers.emptyToNull;
+import static gov.va.api.health.dataquery.service.controller.Transformers.ifPresent;
 import static gov.va.api.health.dataquery.service.controller.Transformers.isBlank;
 import static java.util.Collections.singletonList;
