+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/practitioner/DatamartPractitionerSamples.java
@@ -1,0 +1,69 @@
+package gov.va.api.health.dataquery.service.controller.practitioner;
+
+import gov.va.api.health.dataquery.service.controller.datamart.DatamartCoding;
+import gov.va.api.health.dataquery.service.controller.datamart.DatamartReference;
+import java.time.LocalDate;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor(staticName = "create")
+final class DatamartPractitionerSamples {
+  @AllArgsConstructor(staticName = "create")
+  static class Datamart {
+    public DatamartPractitioner practitioner() {
+      return practitioner("1234");
+    }
+
+    public DatamartPractitioner practitioner(String cdwId) {
+      return DatamartPractitioner.builder()
+          .cdwId(cdwId)
+          .active(true)
+          .address(
+              List.of(
+                  DatamartPractitioner.Address.builder()
+                      .line1("111 MacGyver Viaduct")
+                      .city("Anchorage")
+                      .state("Alaska")
+                      .postalCode("99501")
+                      .build()))
+          .name(DatamartPractitioner.Name.builder().family("Joe").given("Johnson").build())
+          .birthDate(Optional.of(LocalDate.parse("1970-11-14")))
+          .gender(DatamartPractitioner.Gender.male)
+          .npi(Optional.of("1234567"))
+          .practitionerRole(
+              Optional.of(
+                  DatamartPractitioner.PractitionerRole.builder()
+                      .healthCareService(Optional.of("medical"))
+                      .location(
+                          Collections.singletonList(
+                              DatamartReference.builder()
+                                  .display(Optional.of("test"))
+                                  .type(Optional.of("Location"))
+                                  .build()))
+                      .managingOrganization(
+                          Optional.of(
+                              DatamartReference.builder()
+                                  .display(Optional.of("test"))
+                                  .type(Optional.of("Organization"))
+                                  .build()))
+                      .role(
+                          Optional.of(
+                              DatamartCoding.builder()
+                                  .system(Optional.of("test"))
+                                  .display(Optional.of("test"))
+                                  .code(Optional.of("test"))
+                                  .build()))
+                      .build()))
+          .telecom(
+              List.of(
+                  DatamartPractitioner.Telecom.builder()
+                      .use(DatamartPractitioner.Telecom.Use.mobile)
+                      .system(DatamartPractitioner.Telecom.System.phone)
+                      .value("123-456-1234")
+                      .build()))
+          .build();
+    }
+  }
+}
