+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/practitioner/Dstu2PractitionerTransformerTest.java
@@ -81,7 +81,7 @@ public class Dstu2PractitionerTransformerTest {
 
   @Test
   public void practitioner() {
-    assertThat(tx(Dstu2PractitionerSamples.Datamart.create().practitioner()).toFhir())
+    assertThat(tx(DatamartPractitionerSamples.Datamart.create().practitioner()).toFhir())
         .isEqualTo(Dstu2PractitionerSamples.Dstu2.create().practitioner());
   }
