+++ data-query/src/test/java/gov/va/api/health/dataquery/tools/minimart/transformers/F2DDiagnosticReportTransformer.java
@@ -3,7 +3,7 @@ package gov.va.api.health.dataquery.tools.minimart.transformers;
 import static java.util.Arrays.asList;
 
 import gov.va.api.health.argonaut.api.resources.DiagnosticReport;
-import gov.va.api.health.dataquery.service.controller.Dstu2Transformers;
+import gov.va.api.health.dataquery.service.controller.Transformers;
 import gov.va.api.health.dataquery.service.controller.diagnosticreport.DatamartDiagnosticReports;
 import gov.va.api.health.dataquery.tools.minimart.FhirToDatamartUtils;
 import gov.va.api.health.dstu2.api.elements.Reference;
@@ -30,7 +30,7 @@ public class F2DDiagnosticReportTransformer {
         diagnosticReport.performer() != null && diagnosticReport.performer().reference() != null
             ? fauxIds.unmaskByReference(diagnosticReport.performer().reference())
             : null;
-    return Dstu2Transformers.emptyToNull(
+    return Transformers.emptyToNull(
         asList(
             DatamartDiagnosticReports.DiagnosticReport.builder()
                 .identifier(fauxIds.unmask("DiagnosticReport", diagnosticReport.id()))
